Setup webpack config to output the additional server-bundle in a separate config
