+++ config/webpack/config/loaders/typescript.js
